+++ test/Fixture/Classes/FinalRuleWithAttributes/Failure/NonFinalClassWithUnqualifiedDoctrineOrmMappingEntityAttribute.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ergebnis\PHPStan\Rules\Test\Fixture\Classes\FinalRuleWithAttributes\Failure;
+
+#[Doctrine\ORM\Mapping\Entity]
+class NonFinalClassWithUnqualifiedDoctrineOrmMappingEntityAttribute
+{
+}